// lib/types/article.ts
export interface Article {
  $id: string
  title?: string
  articleTitle: string
  introductionSubheading: string
  introductionContent: string
  pexelImgLink: string
  contentOneSubheadingTitle: string
  contentOneParagraph: string
  contentTwoSubheadingTitle: string
  contentTwoParagraph: string
  pexelImgLink2: string
  contentThreeSubheadingTitle?: string
  contentThreeParagraph?: string
  conclusionSubheading: string
  conclusionParagraph: string
  extraSubheading?: string
  extraContentParagraph?: string
}